fix 03: Change OrderDetail model orderLineNumber type from short to int

+++ src/main/java/hu/uni/eku/tzs/model/OrderDetail.java
@@ -21,5 +21,5 @@ public class OrderDetail {
 
     private Double priceEach;
 
-    private short orderLineNumber;
+    private int orderLineNumber;
 }
